gualify route is handling the filter logic
