+++ views/qualify.ejs
@@ -28,7 +28,7 @@
                 </div>
                 <div class="col-xs-12 col-sm-7 pull-left">
                     <p>Fill the below info to check if you are qualified to join the program!</p>
-                    <form action="/decision" method="POST">
+                    <form action="/qualify" method="POST">
                         <dl>
                             <dt><label for="first_name">First Name</label></dt>
                             <dd><input type="text" id="first_name" name="first_name" /></dd>
